Set GIT_MERGE_AUTOEDIT=no
The default behavior is to open an editor for merges. It is
intentionally not exposed as a `git config`, because the idea
is to be intentional about merges, and not pollute git history.

I don't think most people think much of merging w/o editing the
message though. So I will disable.

For background:
   https://github.com/git/git/commit/f8246281af9adb0fdddbcc90d2e19cb5cd5217e5